Move notification components to separate files
